move build.json artifact to workspace root
