+++ app/ide-desktop/lib/content-config/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["../types", "../../build.json", ".", "src/config.json"]
+  "include": ["../types", "../../../../build.json", ".", "src/config.json"]
 }
