+++ app/ide-desktop/lib/types/globals.d.ts
@@ -3,7 +3,7 @@
  * monkeypatching on `window` and generated code. */
 // This file is being imported for its types.
 // eslint-disable-next-line no-restricted-syntax
-import * as buildJson from './build.json' assert { type: 'json' }
+import * as buildJson from './../../build.json' assert { type: 'json' }
 
 // =============
 // === Types ===
